Restructuring files.
