+++ src/components/CompHexGrid.ts
@@ -1,5 +1,5 @@
-import HexGrid from '../../math/hexgrid/HexGrid';
-import Component from '../core/Component';
+import HexGrid from '../math/hexgrid/HexGrid';
+import Component from '../engine/core/Component';
 
 export default class CompHexGrid extends Component {
   constructor(public hexgrid: HexGrid) {
